Update running-command-line-blast.md
I think it's confusing to teach that .faa is fasta for proteins, and then create a subset file from that that we name ".fa". And it's pretty solid convention that amino acid fastas are .faa. So i changed all the ".fa" extensions to ".faa". 

Also, previously stated 4 files and a lost and found directory being present. But we just created this "~/blast" directory and it's empty other than the 3 files we download. 

Last, added the -n flag explicitly to the 'head' commands that are run. And corrected the number of lines being pulled by the previously "head 500" command. That stopped in the middle of sequence. line 498 stops at end, and holds 96 sequences.